+++ cpld.cpp
@@ -51,7 +51,8 @@ ipmi_ret_t CpldVersion(const uint8_t* reqBuf, uint8_t* replyBuf,
 {
     if ((*dataLen) < sizeof(struct CpldRequest))
     {
-        fprintf(stderr, "Invalid command length: %lu\n", (*dataLen));
+        fprintf(stderr, "Invalid command length: %u\n",
+                static_cast<uint32_t>(*dataLen));
         return IPMI_CC_INVALID;
     }
